packages/story-api/pyproject.toml
[project]
name = "story-api"
version = "0.1.0"
description = "Python API for Store E Tails"
authors = [
    {name = "Antonio Reid", email = "read.musik@gmail.com"},
]
dependencies = ["google-genai>=1.5.0"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
